# Refactor & speed up iterative pipeline