--- a/src/entity_factuality.py
+++ b/src/entity_factuality.py
@@ -1,4 +1,4 @@
-from typing import List
+from typing import List, Tuple
 from compute_probs import compute_probs_for_summary
 from src.entity_utils import (
     MarkedEntity,
@@ -12,7 +12,10 @@ from src.generation_utils import (
     load_posterior_model_and_tokenizer,
     load_prior_model_and_tokenizer,
 )
-from src.prob_computation_utils import build_masked_inputs_and_targets_for_inference
+from src.prob_computation_utils import (
+    InferenceInput,
+    build_masked_inputs_and_targets_for_inference,
+)
 import pandas as pd
 
 
@@ -53,33 +56,19 @@ class EntityFactualityClassifier:
                 posterior_model_path
             )
 
-    def extract_features(
-        self,
-        gen_summary: str,
-        source: str,
-        entities: List[MarkedEntity],
-    ):
-        features = np.array(
-            list(
-                zip(
-                    [0] * len(entities),  # priors
-                    [0] * len(entities),  # posteriors
-                    [1.0 if ent["in_source"] else 0.0 for ent in entities],
-                )
-            )
-        )
+    def extract_features(self, ents_to_classify: InferenceInput):
+        features = []
+        for (_, _, ents) in ents_to_classify:
+            for ent in ents:
+                # prior / posterior / in_source
+                features.append([0, 0, 1.0 if ent["in_source"] else 0.0])
+        features = np.array(features)
         (
             inputs,
             targets,
             masked_entities,
             sources,
-        ) = build_masked_inputs_and_targets_for_inference(
-            {
-                "source": source,
-                "prediction": gen_summary,
-                "entities": entities,
-            }
-        )
+        ) = build_masked_inputs_and_targets_for_inference(ents_to_classify)
 
         entity_probs = compute_probs_for_summary(
             masked_inputs=inputs,
@@ -106,21 +95,33 @@ class EntityFactualityClassifier:
         sources_by_id,
     ) -> MarkedEntityLookup:
         classified_entities: MarkedEntityLookup = {}
+        # Build a list of (sum, source, ents[]) to enable
+        # batching across summaries when extracting feaatures
+        ents_to_classify: InferenceInput = []
         for sum_id, summary in gen_summaries_by_id.items():
             updated_entities = [x.copy() for x in marked_entities[sum_id]]
-            entities_to_classify = []
             for ent in updated_entities:
                 if ent["in_source"]:
                     ent["predicted_label"] = ANNOTATION_LABELS["Non-hallucinated"]
-                else:
-                    entities_to_classify.append(ent)
-            if len(entities_to_classify) > 0:
-                features = self.extract_features(
-                    summary, sources_by_id[sum_id], entities_to_classify
-                )
-                predictions = self.clf.predict(features)
-                for entitiy, pred in zip(entities_to_classify, predictions):
-                    entitiy["predicted_label"] = self.label_mapping[pred]
 
+            ents_without_prediction = [ent for ent in updated_entities if not ent["in_source"]]
+            if len(ents_without_prediction) > 0:
+                ents_to_classify.append(
+                    (
+                        summary,
+                        sources_by_id[sum_id],
+                        ents_without_prediction,
+                    )
+                )
             classified_entities[sum_id] = updated_entities
+
+        if len(ents_to_classify) > 0:
+            features = self.extract_features(ents_to_classify)
+            predictions = self.clf.predict(features)
+            idx = 0
+            for (_, _, ents) in ents_to_classify:
+                for entity in ents:
+                    entity["predicted_label"] = self.label_mapping[predictions[idx]]
+                    idx += 1
+
         return classified_entities