**Commit-by-commit:**
1. Parameterized prior/posterior model in iterative constraints pipeline
2. Fixes #36, predicted label is always set by classifier (validated in debugging mode)

### Before
<img width="493" alt="Screenshot 2022-06-03 at 12 51 32" src="https://user-images.githubusercontent.com/1349225/171910291-0a1c755f-de8a-40b0-a042-54f7ed7f1e6a.png">

### After
<img width="547" alt="Screenshot 2022-06-03 at 12 51 36" src="https://user-images.githubusercontent.com/1349225/171910278-4cfdf9ae-403c-49b3-8f45-c383fcf983ee.png">

3. Improved GPU utilization in entity factuality classifier

      Added support for batching across summaries.


This leads to a big (almost 2x) speedup! On a GPU for 100 summaries:

### Before 
```
[Timer]: Iteration 0, 100 docs...
Batch 1/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.445640325546265s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 11.80551815032959s

Batch 2/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.56836199760437s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 11.701605081558228s

Batch 3/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.0058913230896s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 11.637496948242188s

Batch 4/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.922004461288452s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 11.991807222366333s

Batch 5/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.5374414920806885s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 10.295574426651001s

[Timer]: Iteration 0, 100 docs took 81.04801821708679s

[Timer]: Iteration 1, 52 docs...
Batch 1/3
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.236845254898071s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 12.414349794387817s

Batch 2/3
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.938957691192627s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 11.896624326705933s

Batch 3/3
[Timer]: Generating 12 summaries...
[Timer]: Generating 12 summaries took 2.9869933128356934s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 6.171367406845093s

[Timer]: Iteration 1, 52 docs took 43.77115321159363s


````

### After
```
[Timer]: Iteration 0, 100 docs...
Batch 1/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.587785005569458s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 4.194167613983154s

Batch 2/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.438981294631958s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 3.693527936935425s

Batch 3/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.042171478271484s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 3.575554132461548s

Batch 4/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.238194704055786s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 4.258793354034424s

Batch 5/5
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 4.689017057418823s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 3.111659526824951s

[Timer]: Iteration 0, 100 docs took 43.06087040901184s

[Timer]: Iteration 1, 52 docs...
Batch 1/3
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.484066724777222s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 4.6540467739105225s

        
Batch 2/3
[Timer]: Generating 20 summaries...
[Timer]: Generating 20 summaries took 5.036847114562988s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 4.351071834564209s

        
Batch 3/3
[Timer]: Generating 12 summaries...
[Timer]: Generating 12 summaries took 3.036210775375366s
[Timer]: Detecting & classifying entities...
[Timer]: Detecting & classifying entities took 2.3195996284484863s

[Timer]: Iteration 1, 52 docs took 25.090806245803833s
```